+++ shopify/product-bundles-public-template/api/lib/bundle/syncBundleComponentQuantities.ts
